gpet/requirements.txt
aredis
asyncpg
aio-pika
sanic
sanic-jwt
pytest-sanic
qrcode
pillow
aiohttp
APScheduler
Sanic-Cors
aliyun-python-sdk-core-v3
aliyun-python-sdk-green
xlsxwriter
requests
ujson
